@@ -16,26 +16,23 @@
     "check-ci": "yarn run lint",
     "clean": "rimraf dist",
     "format": "yarn run lint --fix && yarn run prettier-all",
-    "lint": "eslint '*.js' src scripts 'static-config-*/*.js'",
-    "prettier-all": "prettier --write '{src,static-config-helpers,static-config-parts,static.config,scripts}/**/*.{js,jsx}'",
+    "lint": "eslint '*.js'",
+    "prettier-all": "prettier --write '{src,static.config,scripts,content}/**/*.{js,jsx,md}'",
     "preversion": "yarn run lint && yarn run build",
-    "prod:build": "react-static build",
-    "prod:deploy": "node scripts/deploy/aws.js",
-    "prod:serve": "serve -l 3000",
-    "stage:build": "react-static build --staging",
-    "stage:deploy": "node scripts/deploy/surge.js",
-    "stage:serve": "serve -l 3000 -c serve-staging.json",
+    "build": "react-static build",
+    "serve": "serve -l 4000",
+    "deploy:prod": "node scripts/deploy/aws.js",
+    "deploy:stage": "node scripts/deploy/surge.js",
     "start": "react-static start",
     "test-local": "node test/spectacle-integration.js"
   },
   "dependencies": {
     "@babel/node": "^7.2.0",
+    "@mdx-js/react": "^1.5.8",
     "array-shuffle": "^1.0.1",
     "axios": "^0.19.1",
-    "chokidar": "2.0.3",
-    "classnames": "^2.2.6",
     "exenv": "^1.2.2",
-    "formidable-oss-badges": "0.1.1",
+    "formidable-oss-badges": "^0.3.4",
     "gray-matter": "^3.1.1",
     "history": "^4.7.2",
     "htmr": "0.6.3",
@@ -44,26 +41,20 @@
     "mdast": "^3.0.0",
     "prismjs": "^1.15.0",
     "prop-types": "^15.6.2",
-    "raw-loader": "^1.0.0",
-    "react": "^16.7.0",
+    "react": "^16.8.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.8.0",
     "react-ga": "^2.4.1",
-    "react-hot-loader": "^4.0.0-beta.21",
-    "react-router": "^4.2.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-router-ga": "^1.0.0",
-    "react-side-effect": "^1.1.5",
-    "react-static": "^5.9.12",
+    "react-static": "^7.2",
+    "react-static-plugin-md-pages": "^0.1.x",
     "styled-components": "^4.1.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.2.0",
     "@babel/core": "^7.2.0",
-    "anchorate": "^1.1.0",
-    "babel-preset-stage-1": "^6.24.1",
-    "chalk": "^2.4.1",
-    "change-case": "^3.0.2",
-    "config": "^3.0.0",
     "eslint-config-formidable": "^4.0.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-config-react-tools": "1.x.x",
@@ -72,33 +63,16 @@
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-react-hooks": "^2.3.0",
     "execa": "^4.0.0",
-    "folder-reader": "^2.0.0",
-    "gatsby-remark-prismjs": "^3.2.2",
-    "github-slugger": "^1.2.1",
-    "js-yaml": "^3.12.0",
-    "json-to-frontmatter-markdown": "^1.0.0",
-    "lodash": "^4.17.11",
-    "node-dir": "^0.1.17",
     "prettier": "^1.15.1",
     "puppeteer": "^1.13.0",
     "react-document-meta": "^2.0.3",
     "react-router-scroll": "^0.2.0",
-    "react-static-plugin-styled-components": "^6.0.0-beta.28",
-    "remark": "^10.0.1",
-    "remark-frontmatter": "^1.3.0",
-    "remark-html": "^8.0.0",
-    "remark-kbd": "^1.0.16",
-    "remark-slug": "^5.1.1",
-    "remark-toc": "^5.1.0",
+    "react-static-plugin-react-router": "^7.2.3",
+    "react-static-plugin-sitemap": "^7.2.3",
+    "react-static-plugin-styled-components": "^7.2.2",
     "rimraf": "^3.0.0",
     "serve": "^11.3.0",
     "spectacle": "^5.0.0",
-    "style-loader": "^0.23.1",
-    "surge": "^0.21.3",
-    "to-vfile": "^5.0.2",
-    "toml-js": "0.0.8",
-    "unist-util-select": "^1.5.0",
-    "unist-util-visit": "^1.4.0",
-    "url-loader": "^1.1.2"
+    "surge": "^0.21.3"
   }
 }
