@@ -1,112 +1,132 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { Wrapper } from './wrapper';
 import styled from 'styled-components';
-import logoFormidableWhite from '../static/svgs/logo_formidable_white.png';
 
-const Container = styled.footer`
-  background: #1f1f1f;
-  color: white;
+import { Wrapper } from './wrapper';
+import { theme } from '../theme';
+import logoFormidableWhite from '../assets/logo_formidable_white.png';
+
+const FooterContainer = styled.div`
   display: flex;
   flex-direction: column;
-  height: auto;
-  padding: 9rem 6rem;
-  @media (max-width: 768px) {
-    padding: 6rem 4rem 8rem 3.5rem;
-  }
-  align-items: left;
-`;
 
-const FooterDescription = styled.p`
-  flex: 2;
-  font-size: 1.4rem;
-  letter-spacing: 0.05rem;
-  line-height: 1.6;
-  margin: 2rem 0 0;
-  max-width: 56rem;
-  text-align: left;
-  @media (min-width: 768px) {
-    font-size: 1.5rem;
-    margin: 0;
-    min-width: auto;
+  @media ${p => p.theme.media.sm} {
+    flex-direction: row;
   }
-  & a {
-    color: white;
-    letter-spacing: 0.05em;
-    transition: opacity 0.4s;
-  }
-  & a:hover {
-    opacity: 0.7;
-  }
-  & a:visited {
-    color: white;
+
+  > * {
+    flex-basis: 50%;
   }
 `;
 
 const FooterLeft = styled.div`
   display: flex;
-  flex: 1;
   padding: 0;
   text-align: left;
 `;
 
 const FooterLogo = styled.img`
-  width: 7rem;
-  margin-right: 2.7rem;
+  width: 10rem;
+  margin-right: 3rem;
 `;
 
 const FooterLinks = styled.ul`
   font-size: 1.4rem;
   list-style: none;
-  padding: 0px 8px;
+  padding: 0 1rem;
   text-transform: uppercase;
-  & li {
-    margin-bottom: 1.4rem;
+
+  > * {
+    margin-top: 0;
+    margin-bottom: 0;
   }
+
+  > * + * {
+    margin-top: 1.5rem;
+  }
+
   & a {
-    color: white;
-    letter-spacing: 0.05em;
-    transition: opacity 0.4s;
+    color: ${p => p.theme.colors.textLight};
+    letter-spacing: 0.1rem;
+    transition: opacity 0.4s ease-out;
   }
+
   & a:hover {
     opacity: 0.7;
   }
+
   & a:visited {
-    color: white;
+    color: ${p => p.theme.colors.textLight};
   }
 `;
 
-export const Footer = props => (
-  <Container>
-    <Wrapper noPadding noMargin={props.articleFooter}>
-      <FooterLeft>
-        <a href="https://formidable.com" title="Formidable">
-          <FooterLogo src={logoFormidableWhite} alt="Formidable Logo" />
-        </a>
-        <FooterLinks>
-          <li>
-            <a href="https://formidable.com/contact/" title="Contact">
-              Contact
-            </a>
-          </li>
-          <li>
-            <a href="https://formidable.com/careers/" title="Careers">
-              Careers
-            </a>
-          </li>
-        </FooterLinks>
-      </FooterLeft>
-      <FooterDescription>
-        Formidable is a Seattle, Denver, and London-based engineering
-        consultancy and open source software organization, specializing in
-        React.js, React Native, GraphQL, Node.js, and the extended JavaScript
-        ecosystem. For more information about Formidable, please visit{' '}
-        <a href="https://www.formidable.com">formidable.com</a>.
-      </FooterDescription>
-    </Wrapper>
-  </Container>
-);
+const FooterDescription = styled.p`
+  font-size: 1.4rem;
+  letter-spacing: 0.05rem;
+  line-height: 1.6;
+  max-width: 56rem;
+  text-align: left;
+  color: ${p => p.theme.colors.textLight};
+  margin: 0;
 
-Footer.propTypes = {
-  articleFooter: PropTypes.bool
+  & a {
+    color: ${p => p.theme.colors.textLight};
+    letter-spacing: 0.1rem;
+    transition: opacity 0.3s ease-out;
+  }
+
+  & a:hover {
+    opacity: 0.7;
+  }
+
+  & a:visited {
+    color: ${p => p.theme.colors.textLight};
+  }
+`;
+
+export const Footer = () => {
+  return (
+    <Wrapper background={theme.colors.bgDark}>
+      <FooterContainer>
+        <FooterLeft>
+          <a
+            href="https://formidable.com"
+            title="Formidable"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <FooterLogo src={logoFormidableWhite} alt="Formidable Logo" />
+          </a>
+          <FooterLinks>
+            <li>
+              <a
+                href="https://formidable.com/contact/"
+                title="Contact"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                Contact
+              </a>
+            </li>
+            <li>
+              <a
+                href="https://formidable.com/careers/"
+                title="Careers"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                Careers
+              </a>
+            </li>
+          </FooterLinks>
+        </FooterLeft>
+        <FooterDescription>
+          Formidable is a Seattle, Denver, and London-based engineering
+          consultancy and open source software organization, specializing in
+          React.js, React Native, GraphQL, Node.js, and the extended JavaScript
+          ecosystem. For more information about Formidable, please visit{' '}
+          <a href="https://www.formidable.com">formidable.com</a>.
+        </FooterDescription>
+      </FooterContainer>
+    </Wrapper>
+  );
 };
