@@ -1,104 +1,118 @@
 import styled from 'styled-components';
-import { Link } from 'react-static';
-import sidebarBackground from '../static/svgs/sidebar-background.svg';
-import collapsedSidebarBackground from '../static/svgs/collapsed-sidebar-background.svg';
+import { NavLink } from 'react-router-dom';
+import SvgChevron from '../assets/chevron';
 
-const sidebarZIndex = 900;
+export const SidebarContainer = styled.div`
+  display: ${p => (p.hidden ? 'none' : 'block')};
+  position: absolute;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  min-height: 100%;
 
-export const Navigation = styled.div`
-  align-items: center;
-  background: rgba(0, 0, 0, 0.3);
-  display: flex;
-  flex-direction: row;
-  height: 6rem;
-  width: 100%;
-
-  & img {
-    margin-left: auto;
-    @media (min-width: 768px) {
-      margin-left: 0;
-    }
+  @media ${p => p.theme.media.sm} {
+    display: block;
+    position: relative;
+    width: ${p => p.theme.layout.sidebar};
+    margin-left: calc(2 * ${p => p.theme.layout.stripes});
   }
 `;
-export const SidebarContainer = styled.div`
-  width: 24rem;
-  min-width: 24rem;
-  min-height: 100vh;
 
-  @media (max-width: 768px) {
-    min-width: 2.5rem;
-    width: 2.5rem;
-  }
+export const SidebarStripes = styled.div`
+  border-left: ${p => p.theme.layout.stripes} solid
+    ${p => p.theme.colors.accentLight};
+  border-right: ${p => p.theme.layout.stripes} solid
+    ${p => p.theme.colors.accentMedium};
+  position: absolute;
+  height: 100%;
+  width: 0;
+  left: 0;
+  top: 0;
+  bottom: 0;
 `;
 
 export const SidebarWrapper = styled.aside`
-  background-image: url(${sidebarBackground});
-  background-repeat: repeat-y;
-  background-size: 100%;
-  min-height: 100vh;
-  padding-top: 18rem;
-  min-width: 24rem;
-  width: 24rem;
-  z-index: ${sidebarZIndex};
   position: fixed;
-  overflow-y: scroll;
-  top: 0;
   bottom: 0;
+  top: ${p => p.theme.layout.header};
+  -webkit-overflow-scrolling: touch;
+  overflow-y: scroll;
 
-  @media (max-width: 768px) {
-    background-image: ${props =>
-      props.overlay
-        ? `url(${sidebarBackground})`
-        : `url(${collapsedSidebarBackground})`};
-    min-width: ${props => (props.overlay ? '24rem' : '2.5rem')};
-    width: ${props => (props.overlay ? '24rem' : '2.5rem')};
+  display: flex;
+  flex-direction: column;
+  z-index: 1;
+  overflow-y: scroll;
+  min-height: 100%;
+  line-height: ${p => p.theme.lineHeights.body};
+  font-size: ${p => p.theme.fontSizes.small};
+
+  padding: ${p => p.theme.spacing.sm} ${p => p.theme.spacing.md};
+  background-color: ${p => p.theme.colors.accentDark};
+  border-right: 1px solid ${p => p.theme.colors.border};
+  border-top: 1px solid ${p => p.theme.colors.border};
+
+  @media ${({ theme }) => theme.media.sm} {
+    border: none;
+    padding-top: ${p => p.theme.spacing.md};
+    width: ${p => p.theme.layout.sidebar};
   }
 `;
 
-export const SidebarNavItem = styled(Link)`
-  color: white;
-  font-size: 1.6rem;
+export const SidebarNavItem = styled(NavLink).attrs(() => ({
+  activeClassName: 'active'
+}))`
+  display: block;
+  margin: ${p => p.theme.spacing.xs} 0;
+  color: ${p => p.theme.colors.textLight};
+  font-weight: ${p => p.theme.fontWeights.heading};
+  text-decoration: none;
+  width: 100%;
+  &:hover {
+    color: ${p => p.theme.colors.accentLight};
+  }
+  &.active {
+    color: ${p => p.theme.colors.accentLight};
+  }
+`;
+
+export const ChevronItem = styled(SvgChevron).attrs(() => ({
+  'aria-hidden': 'true'
+}))`
   display: inline-block;
-  margin: 0.25rem 0;
-  padding: 0.25rem;
-  font-weight: bold;
+  color: inherit;
+  vertical-align: baseline;
+  margin-top: 0.08em;
+  margin-left: 0.3em;
+  padding: 0.08em;
+  width: 1em;
+  height: 1em;
+  position: relative;
+  top: 0.16em;
 
-  &:hover {
-    color: rgba(200, 200, 200, 0.8);
-  }
-
-  /* For selected state (need to break out of container) */
-  &::after {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 3rem;
-    background: ${({ isSelected }) =>
-      isSelected ? 'rgba(0, 0, 0, 0.2)' : 'transparent'};
+  ${SidebarNavItem}.active & {
+    transform: rotate(180deg);
   }
 `;
 
-export const SidebarNavSubItem = styled(Link)`
-  color:  ${({ isSelected }) =>
-    isSelected ? 'rgba(255, 255, 255, 0.8)' : 'white'};
-  }
-  margin-left: 3rem;
-  margin-top: 1rem;
-  font-size: 1.4rem;
-  &:hover {
-    color: rgba(200, 200, 200, 0.8);
-  }
+export const SidebarNavSubItemWrapper = styled.div`
+  padding-left: ${p => p.theme.spacing.sm};
+  margin-bottom: ${p => p.theme.spacing.xs};
+`;
 
-  /* For selected state (need to break out of container) */
-  &::after {
-    content: "";
-    position: absolute;
-    left: 0;
-    width: 100%;
-    height: 2rem;
-    background: ${({ isSelected }) =>
-      isSelected ? 'rgba(0, 0, 0, 0.2)' : 'transparent'};
+export const SidebarNavSubItem = styled(NavLink).attrs(() => ({}))`
+  display: block;
+  color: ${p => p.theme.colors.textLight};
+  font-weight: ${p => p.theme.fontWeights.body};
+  text-decoration: none;
+  margin-top: ${p => p.theme.spacing.xs};
+  &:first-child {
+    margin-top: 0;
+  }
+  &:hover {
+    color: ${p => p.theme.colors.accentLight};
+  }
+  &.active {
+    color: ${p => p.theme.colors.accentLight};
+    font-weight: ${p => p.theme.fontWeights.heading};
   }
 `;
