@@ -1,18 +1,8 @@
 import styled from 'styled-components';
 
 export const SecondaryTitle = styled.h3`
-  color: #4e4e4e;
+  color: ${({ theme }) => theme.colors.textDark};
   font-size: 2rem;
   line-height: 2.4rem;
-  margin: 2.3rem auto 1rem;
   text-align: center;
-  max-width: 26rem;
-  @media (min-width: 768px) {
-    font-size: 2.2rem;
-    line-height: 2.6rem;
-    margin-top: 3.1rem;
-  }
-  @media (min-width: 1024px) {
-    margin-top: 7.5rem;
-  }
 `;
