@@ -15,52 +15,50 @@ const content = {
       icon: require('../../static/svgs/code-preview.svg')
     },
     {
-      title: 'Auto-Size Text, Image Dimming, and More',
+      title: '... and More!',
       description:
-        "On top of all of Spectacle's helpful features, you can also make your presentation look amazing with auto-formatting, easy themeing abilities, image dimming, and lots of other fun touches",
+        'Polish off your presentation with auto-formatting, easy themeing, image dimming, and other fun touches available out of the box.',
       icon: require('../../static/svgs/amazing.svg')
     }
   ],
+  preview: {
+    bgMp4: require('../../assets/demo-presentation.mp4'),
+    bgStill: require('../../assets/demo-still.png'),
+    bgWebm: require('../../assets/demo-presentation.webm'),
+    demoUrl:
+      'https://raw.githack.com/FormidableLabs/spectacle/master/examples/one-page.html'
+  },
   getStarted: {
     description:
       'There are several flexible options for getting started with Spectacle, using either JSX or MDX syntax - dive into the documentation to see all the ways you can get a presentation up and running.',
-    link: '/docs'
+    link: '/docs/basic-concepts/'
   },
   oss: [
+    {
+      title: 'Renature',
+      description:
+        'A collection of animations that can be used with many inline style libraries, such as Radium or Aphrodite.',
+      link: 'https://formidable.com/open-source/renature'
+    },
     {
       title: 'Victory',
       description:
         'An ecosystem of modular data visualization components for React. Friendly and flexible.',
-      logo: require('../../static/svgs/logo_victory.svg'),
-      link: 'https://formidable.com/open-source/victory',
-      hasOwnLogo: true
+      link: 'https://formidable.com/open-source/victory'
     },
     {
-      title: 'Development Dashboards',
+      title: 'urql',
       description:
-        'Dashboards to organize and intuitively display your dev server and tooling output.',
-      abbreviation: 'Dd',
-      color: '#8bd48b',
-      number: '17',
-      link: 'https://formidable.com/open-source/development-dashboards/'
+        'Universal React Query Library is a blazing-fast GraphQL client, exposed as a set of ReactJS components.',
+      link: 'https://formidable.com/open-source/urql/'
     },
     {
-      title: 'React Animations',
+      title: 'Runpkg',
       description:
-        'A collection of animations that can be used with many inline style libraries, such as Radium or Aphrodite.',
-      abbreviation: 'Ra',
-      color: '#86b9e6',
-      number: '03',
-      link: 'https://formidable.com/open-source/react-animations'
-    },
-    {
-      title: 'Enzyme Matchers',
-      description:
-        'Run common assertions on your React components using Enzyme in a Jest or Jasmine environment.',
-      abbreviation: 'Em',
-      color: '#e48055',
-      number: '09',
-      link: 'https://formidable.com/open-source/jest-enzyme/'
+        'The online package explorer. Runpkg turns any npm package into an interactive and informative browsing experience.',
+      abbreviation: 'Rp',
+      color: '#80eac7',
+      link: 'https://www.runpkg.com/'
     }
   ]
 };
