@@ -9,109 +9,146 @@ export default function SpectacleLogo(props) {
       height={props.size}
       viewBox="0 0 601 595"
     >
-      <g fill="none" fillRule="evenodd">
-        <path
-          fill="#202020"
-          d="M342.277 19.4l195.971 87.165C575.801 123.268 600 160.515 600 201.616v211.961c0 41.1-24.199 78.348-61.752 95.051l-195.971 87.165a104.029 104.029 0 01-84.554 0L61.752 508.628C24.199 491.925 0 454.678 0 413.578V201.615c0-41.1 24.199-78.348 61.752-95.05L257.723 19.4a104.029 104.029 0 0184.554 0z"
-          transform="translate(.667 -10)"
-        />
-        <path
-          stroke="#FC6986"
-          strokeWidth="24.247"
-          d="M332.269 53.303a79.877 79.877 0 00-64.741 0L76.38 138.036c-24.842 11.012-40.858 35.632-40.858 62.805v213.702c0 27.174 16.016 51.793 40.858 62.805l191.148 84.734a79.877 79.877 0 0064.74 0l191.148-84.734c24.842-11.012 40.859-35.631 40.859-62.805V200.841c0-27.173-16.017-51.793-40.859-62.805L332.27 53.303z"
-          opacity=".499"
-          transform="translate(.667 -10)"
-        />
-        <path
-          stroke="#FC6986"
-          strokeWidth="24.007"
-          d="M337.399 30.368a92.026 92.026 0 00-74.798 0l-195.97 87.165c-33.221 14.775-54.628 47.725-54.628 84.083v211.961c0 36.358 21.407 69.308 54.627 84.084l195.971 87.164a92.026 92.026 0 0074.798 0l195.97-87.164c33.221-14.776 54.628-47.726 54.628-84.084v-211.96c0-36.359-21.407-69.309-54.627-84.084L337.399 30.368z"
-          transform="translate(.667 -10)"
-        />
-        <g>
-          <path
-            fill="#F14E74"
-            d="M157.997 216.514L50.114 111.686C36.908 98.854 38.448 77.149 53.33 66.33c14.884-10.82 35.883-5.501 43.882 11.116l65.348 135.75c1.392 2.891-2.266 5.55-4.564 3.318"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FC6986"
-            d="M162.555 213.17L97.09 77.479c-8.014-16.61.855-36.461 18.533-41.483 17.677-5.023 35.578 7.223 37.407 25.59l14.947 150.046c.318 3.196-4.027 4.43-5.422 1.54"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#E13E65"
-            d="M154.775 221.155l-137.206-61.1c-16.795-7.48-22.786-28.39-12.52-43.697 10.267-15.307 31.805-17.577 45.008-4.743L157.924 216.46c2.297 2.233-.226 5.996-3.149 4.694"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FC6986"
-            d="M4.005 118.137c.736-1.507 6.489-12.257 22.256-14.9 7.154-1.198 17.406 1.147 25.97 10.593l99.313 96.342.715 4.147-97.933-76.994s-15.567-12.2-33.372-.767l-.043-.046a12.583 12.583 0 01-6.663 1.913c-7.015 0-12.702-5.715-12.702-12.763 0-2.818.92-5.413 2.46-7.525"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FF92AC"
-            d="M51.446 67.702c1.203-1.166 10.262-9.297 25.98-6.375 7.131 1.326 15.971 7.047 20.814 18.862l60.587 125.147-.736 4.144L92.23 102.992S81.739 86.186 61.12 90.828l-.025-.058a12.569 12.569 0 01-6.913-.486c-6.594-2.406-9.998-9.727-7.605-16.352a12.68 12.68 0 014.868-6.23M113.547 36.788c1.529-.678 12.82-5.18 26.577 3.002 6.24 3.71 12.58 12.138 13.08 24.907l14.01 138.455-2.11 3.635-25.356-122.76s-4.091-19.407-25.04-22.18l-.004-.062a12.587 12.587 0 01-6.323-2.85c-5.367-4.538-6.055-12.589-1.539-17.981a12.619 12.619 0 016.705-4.166"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FF92AC"
-            d="M151.276 53.953l-1.021 133.213s4.055 29.884 9.105 36.032l-23.665-10.511c-8.94-7.605-6.593-24.569-6.593-24.569l22.174-134.165z"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#E13E65"
-            d="M168.916 224.707c-17.639 0-31.938-14.367-31.938-32.09L155.552 12.77c.711-6.88 6.48-12.106 13.364-12.106 6.883 0 12.653 5.226 13.364 12.106l18.574 179.847c0 17.723-14.3 32.09-31.938 32.09"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FF92AC"
-            d="M159.625 31.361L168.931 47.845 178.285 31.572"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FC6986"
-            d="M182.264 14.076c0 7.407-2.816 24.036-13.348 24.036-10.091 0-13.349-16.629-13.349-24.036S161.544.664 168.916.664s13.348 6.005 13.348 13.412M222.789 35.994c-15.437-4.386-31.043 4.398-35.97 18.926l16.575 101.17 37.927-78.611c8.013-16.61-.855-36.463-18.532-41.485M169.523 201.453c-1.096-.137-17.918-2.58-17.507-22.395l3.589-89.137a114.848 114.848 0 00-4.317-36.15l-14.32 138.864a7.42 7.42 0 00.074 2.034c.37 5.013 2.534 18.102 14.922 25.297 4.188 2.432 9.659 4.675 16.396 4.732l.001-.026c.107.003.21.016.318.016 6.397 0 11.583-5.21 11.583-11.64 0-6.14-4.737-11.16-10.74-11.595"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FF92AC"
-            d="M140.853 89.343s2.258 13.802-.282 29.306l2.54-7.94 2.446-18.908-4.704-2.458zM231.569 40.954a12.619 12.619 0 00-6.705-4.166c-1.528-.678-12.82-5.18-26.577 3.001-4.512 2.684-9.065 7.847-11.436 15.322l7.867 48.02 3.945-19.104s4.091-19.406 25.04-22.18l.004-.062a12.592 12.592 0 006.323-2.849c5.368-4.539 6.056-12.59 1.539-17.982"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#F14E74"
-            d="M285.08 66.33c-14.883-10.82-35.883-5.5-43.881 11.116l-37.818 78.56 5.327 32.518s.024.156.058.44l79.532-77.278c13.206-12.832 11.665-34.537-3.218-45.356"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FF92AC"
-            d="M291.833 73.932a12.685 12.685 0 00-4.867-6.23c-1.204-1.166-10.262-9.298-25.981-6.374-7.131 1.325-15.971 7.045-20.813 18.86l-36.77 75.95 2.076 12.667 40.704-65.812s10.49-16.807 31.108-12.165l.025-.058c2.232.468 4.613.353 6.913-.487 6.595-2.405 10-9.726 7.605-16.351"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#E13E65"
-            d="M333.362 116.358c-10.267-15.307-31.805-17.577-45.008-4.743l-79.587 77.358c.325 2.7 1.61 16.966-6.418 23.849l118.493-52.767c16.795-7.48 22.786-28.39 12.52-43.697"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-          <path
-            fill="#FC6986"
-            d="M334.406 118.137c-.735-1.507-6.488-12.257-22.256-14.9-7.154-1.199-17.406 1.147-25.97 10.593l-77.415 75.128c.128 1.057.407 3.946.2 7.476l75.12-59.109s15.567-12.201 33.372-.767l.043-.046a12.58 12.58 0 006.664 1.913c7.015 0 12.702-5.715 12.702-12.763 0-2.818-.92-5.413-2.46-7.525"
-            transform="translate(.667 -10) translate(130 145)"
-          />
-        </g>
-        <text
-          fill="#FC6986"
-          fontFamily="Castledown-Bold, Castledown"
-          fontSize="52.014"
-          fontWeight="bold"
-          letterSpacing="3.853"
-          transform="translate(.667 -10)"
+      <title>Spectacle</title>
+      <g
+        id="Finals"
+        stroke="none"
+        strokeWidth="1"
+        fill="none"
+        fillRule="evenodd"
+      >
+        <g
+          id="The-Good-Badges-for-OSS-usage"
+          transform="translate(-1006.000000, -337.000000)"
         >
-          <tspan x="143" y="472">
-            SPECTACLE
-          </tspan>
-        </text>
+          <g id="Badge-Copy-4" transform="translate(1006.666667, 327.000000)">
+            <path
+              d="M342.276968,19.4007702 L538.247998,106.565184 C575.801346,123.268242 600,160.515702 600,201.616141 L600,413.57732 C600,454.677759 575.801346,491.925219 538.247998,508.628277 L342.276968,595.79269 C315.36338,607.763373 284.63662,607.763373 257.723032,595.79269 L61.7520017,508.628277 C24.1986538,491.925219 2.32856136e-14,454.677759 0,413.57732 L0,201.616141 C2.33883573e-14,160.515702 24.1986538,123.268242 61.7520017,106.565184 L257.723032,19.4007702 C284.63662,7.43008735 315.36338,7.43008735 342.276968,19.4007702 Z"
+              id="Polygon-Copy-4"
+              fill="#202020"
+            ></path>
+            <path
+              d="M332.268654,53.3025417 C311.655676,44.165094 288.140981,44.165094 267.528004,53.3025417 L76.380235,138.035705 C51.5383581,149.047766 35.5217047,173.667505 35.5217047,200.840733 L35.5217047,414.543463 C35.5217047,441.716691 51.5383581,466.336431 76.380235,477.348491 L267.528004,562.081654 C288.140981,571.219102 311.655676,571.219102 332.268654,562.081654 L523.416422,477.348491 C548.258299,466.336431 564.274953,441.716691 564.274953,414.543463 L564.274953,200.840733 C564.274953,173.667505 548.258299,149.047766 523.416422,138.035705 L332.268654,53.3025417 Z"
+              id="Polygon-Copy-5"
+              stroke="#FC6986"
+              strokeWidth="24.2467521"
+              opacity="0.499209449"
+            ></path>
+            <path
+              d="M337.398856,30.3681883 C313.590683,19.7787381 286.409317,19.7787381 262.601144,30.3681883 L66.6301133,117.532602 C33.409844,132.308384 12.0033426,165.25806 12.0033426,201.616141 L12.0033426,413.57732 C12.0033426,449.9354 33.409844,482.885076 66.6301133,497.660859 L262.601144,584.825272 C286.409317,595.414722 313.590683,595.414722 337.398856,584.825272 L533.369887,497.660859 C566.590156,482.885076 587.996657,449.9354 587.996657,413.57732 L587.996657,201.616141 C587.996657,165.25806 566.590156,132.308384 533.369887,117.532602 L337.398856,30.3681883 Z"
+              id="Polygon-Copy"
+              stroke="#FC6986"
+              strokeWidth="24.0066852"
+            ></path>
+            <g id="peacock" transform="translate(130.000000, 145.000000)">
+              <g id="Group-41">
+                <path
+                  d="M157.997195,216.513892 L50.113914,111.685631 C36.9081129,98.854247 38.4478205,77.1487901 53.3314404,66.3299245 C68.2150602,55.5097309 89.2138965,60.8288626 97.213107,77.4462524 L162.56147,213.195992 C163.953154,216.086954 160.294861,218.746188 157.997195,216.513892"
+                  id="Fill-1"
+                  fill="#F14E74"
+                ></path>
+                <path
+                  d="M162.555456,213.170628 L97.0901287,77.4779906 C89.0763801,60.8685685 97.9452281,41.0169395 115.622789,35.9946062 C133.300351,30.9722728 151.201269,43.2180282 153.030415,61.5843374 L167.976813,211.630862 C168.295327,214.82659 163.949784,216.06159 162.555456,213.170628"
+                  id="Fill-3"
+                  fill="#FC6986"
+                ></path>
+                <path
+                  d="M154.775373,221.155235 L17.5689228,160.054606 C0.774215205,152.575552 -5.21742807,131.664877 5.04905146,116.358181 C15.315531,101.050821 36.8535953,98.7813422 50.057414,111.614719 L157.924174,216.460907 C160.22118,218.693866 157.698174,222.456633 154.775373,221.155235"
+                  id="Fill-5"
+                  fill="#E13E65"
+                ></path>
+                <path
+                  d="M4.00509006,118.137444 C4.74058129,116.629549 10.493669,105.879737 26.2614643,103.237103 C33.4148269,102.038622 43.6667684,104.383794 52.2309789,113.829552 L151.544102,210.171503 L152.259108,214.318713 L54.3264351,137.325095 C54.3264351,137.325095 38.7588678,125.12449 20.9537684,136.558199 L20.911476,136.512385 C18.9726339,137.716842 16.6947918,138.424643 14.2477801,138.424643 C7.23252865,138.424643 1.54618363,132.710444 1.54618363,125.662312 C1.54618363,122.844388 2.46604327,120.248896 4.00509006,118.137444"
+                  id="Fill-8"
+                  fill="#FC6986"
+                ></path>
+                <path
+                  d="M51.445662,67.7016376 C52.6490129,66.5363554 61.7075158,58.4046108 77.4264105,61.3274441 C84.5573053,62.6527452 93.3970772,68.3735839 98.2402175,80.1890812 L158.827381,205.336412 L158.091229,209.480302 L92.2287497,102.992095 C92.2287497,102.992095 81.7395743,86.1861349 61.1207088,90.8280086 L61.0962585,90.7702425 C58.8646737,91.2383473 56.483083,91.1221511 54.1827731,90.2835462 C47.589124,87.8779522 44.184586,80.5569253 46.5780713,73.9317478 C47.5355977,71.2831376 49.2814807,69.1584065 51.445662,67.7016376"
+                  id="Fill-10"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M113.547092,36.7877285 C115.075565,36.1098065 126.366975,31.6080323 140.123899,39.7895753 C146.364671,43.4998871 152.704565,51.9284301 153.203484,64.6967366 L167.214162,203.152188 L165.103507,206.78747 L139.747893,84.0271425 C139.747893,84.0271425 135.656764,64.620379 114.70815,61.8469409 L114.704185,61.7845269 C112.448811,61.4518737 110.252911,60.5196478 108.380811,58.9353952 C103.014302,54.3971022 102.325729,46.3463629 106.842425,40.9535296 C108.648443,38.7975914 111.016156,37.4072204 113.547092,36.7877285"
+                  id="Fill-12"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M151.276404,53.9526349 L150.254778,187.165976 C150.254778,187.165976 154.309562,217.05032 159.359539,223.198097 L135.69496,212.686654 C126.754743,205.082108 129.101971,188.118122 129.101971,188.118122 L151.276404,53.9526349 Z"
+                  id="Fill-14"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M168.915836,224.707254 L168.915836,224.707254 C151.277263,224.707254 136.977807,210.340087 136.977807,192.617173 L155.552099,12.7699664 C156.26314,5.89048522 162.032088,0.663646505 168.915836,0.663646505 C175.798924,0.663646505 181.568532,5.89048522 182.279573,12.7699664 L200.853865,192.617173 C200.853865,210.340087 186.554409,224.707254 168.915836,224.707254"
+                  id="Fill-16"
+                  fill="#E13E65"
+                ></path>
+                <polyline
+                  id="Fill-19"
+                  fill="#FF92AC"
+                  points="159.624989 31.3611651 168.931299 47.8450952 178.284527 31.5716462"
+                ></polyline>
+                <path
+                  d="M182.263846,14.0756137 C182.263846,21.4829578 179.448097,38.1116352 168.915968,38.1116352 C158.824606,38.1116352 155.56743,21.4829578 155.56743,14.0756137 C155.56743,6.6689336 161.543875,0.663912097 168.915968,0.663912097 C176.288062,0.663912097 182.263846,6.6689336 182.263846,14.0756137"
+                  id="Fill-21"
+                  fill="#FC6986"
+                ></path>
+                <path
+                  d="M222.788685,35.994075 L222.788685,35.994075 C207.35196,31.608497 191.746065,40.3922685 186.819662,54.920118 L203.394317,156.089857 L241.320685,77.4787874 C249.334434,60.8687013 240.465586,41.0164083 222.788685,35.994075"
+                  id="Fill-24"
+                  fill="#FC6986"
+                ></path>
+                <path
+                  d="M169.522798,201.453134 C168.427161,201.31569 151.604699,198.872249 152.015728,179.057803 L155.604635,89.921346 C156.096944,77.7174212 154.639178,65.5121685 151.287506,53.7703728 L136.968225,192.634835 C136.896196,193.331349 136.928576,194.013255 137.042237,194.669265 C137.411635,199.682303 139.575816,212.770647 151.963523,219.965518 C156.151792,222.397671 161.622711,224.640591 168.359757,224.697693 L168.361079,224.671798 C168.468132,224.675117 168.57188,224.687733 168.678933,224.687733 C175.076319,224.687733 180.262424,219.477494 180.262424,213.048854 C180.262424,206.907053 175.525015,201.888703 169.522798,201.453134"
+                  id="Fill-26"
+                  fill="#FC6986"
+                ></path>
+                <path
+                  d="M140.85298,89.3432199 C140.85298,89.3432199 143.110998,103.144677 140.570811,118.648575 L143.110998,110.708056 L145.557349,91.8012683 L140.85298,89.3432199 Z"
+                  id="Fill-28"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M231.568785,40.9537952 C229.762768,38.797857 227.395054,37.407486 224.864119,36.7879941 C223.335645,36.110072 212.044236,31.6082978 198.287312,39.7891769 C193.774581,42.472978 189.22154,47.6360747 186.851183,55.1111446 L194.717569,103.130069 L198.663318,84.0274081 C198.663318,84.0274081 202.754446,64.6206446 223.70306,61.8478704 L223.707025,61.7847925 C225.962399,61.4521392 228.1583,60.5199134 230.030399,58.9363247 C235.397569,54.3973677 236.086142,46.3459645 231.568785,40.9537952"
+                  id="Fill-30"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M285.080233,66.3295925 L285.080233,66.3295925 C270.196613,55.5093989 249.197116,60.8291946 241.198566,77.4459204 L203.380572,156.005864 L208.708092,188.523547 C208.708092,188.523547 208.731882,188.680246 208.766244,188.964428 L288.297759,111.685963 C301.50356,98.8539151 299.963192,77.149122 285.080233,66.3295925"
+                  id="Fill-32"
+                  fill="#F14E74"
+                ></path>
+                <path
+                  d="M291.833007,73.9316151 C290.876142,71.2836688 289.129598,69.1589376 286.966077,67.7015048 C285.762065,66.5362226 276.703563,58.404478 260.984668,61.3279753 C253.853773,62.6532763 245.014001,68.3734511 240.171522,80.1889484 L203.402247,156.138128 L205.477878,168.804846 L246.182329,102.992626 C246.182329,102.992626 256.671504,86.1860022 277.29037,90.8278758 L277.31482,90.7701097 C279.547065,91.2382145 281.927995,91.1226823 284.228305,90.2834134 C290.822615,87.8778194 294.227153,80.5574565 291.833007,73.9316151"
+                  id="Fill-34"
+                  fill="#FF92AC"
+                ></path>
+                <path
+                  d="M333.361961,116.358181 L333.361961,116.357517 C323.095481,101.050821 301.557417,98.7813422 288.354259,111.614719 L208.767236,188.972861 C209.091698,191.673926 210.37699,205.93884 202.348704,212.821641 L320.842089,160.054606 C337.637458,152.575552 343.62844,131.665541 333.361961,116.358181"
+                  id="Fill-36"
+                  fill="#E13E65"
+                ></path>
+                <path
+                  d="M334.40612,118.137444 C333.670629,116.629549 327.917542,105.879737 312.149746,103.237767 C304.996384,102.037958 294.744442,104.383794 286.180232,113.830216 L208.765319,188.958054 C208.892857,190.015108 209.172384,192.904078 208.964887,196.434452 L284.085436,137.325095 C284.085436,137.325095 299.652343,125.123826 317.457442,136.558199 L317.499735,136.512385 C319.439237,137.717506 321.716419,138.424643 324.164091,138.424643 C331.178682,138.424643 336.865688,132.710444 336.865688,125.662312 C336.865688,122.844388 335.945167,120.248896 334.40612,118.137444"
+                  id="Fill-39"
+                  fill="#FC6986"
+                ></path>
+              </g>
+            </g>
+            <text
+              id="SPECTACLE"
+              fontFamily="Helvetica-Bold, Helvetica"
+              fontSize="52.0144847"
+              fontWeight="bold"
+              line-spacing="40.0111421"
+              letterSpacing="3.85292292"
+              fill="#FC6986"
+            >
+              <tspan x="125.5" y="467">
+                SPECTACLE
+              </tspan>
+            </text>
+          </g>
+        </g>
       </g>
     </svg>
   );
